Add files via upload
Notebook 1 takes an input table is from a segregating population, such as an F2 or the dihaploids produced by a potato haploid induction cross. The table has genomic bins as rows and individual as column headings. Reads mapping in a given bin were standardized by the mean counts obtained in the same bin. The notebook analyzes the dosage variation in each genomic bin, catalogs it into one or more dosage clusters (peaks), and analyzes the correlation between different bins.